+++ meridian/analysis/optimizer_test.py
@@ -885,10 +885,11 @@ class OptimizerAlgorithmTest(parameterized.TestCase):
             c.CHANNEL: self.input_data_media_and_rf.get_all_channels(),
         },
     )
+    optimization_grid = (
+        self.budget_optimizer_media_and_rf.create_optimization_grid()
+    )
 
-    optimization_results = self.budget_optimizer_media_and_rf.optimize()
-
-    actual_data = optimization_results.optimization_grid
+    actual_data = optimization_grid.grid_dataset
     self.assertEqual(actual_data, expected_data)
     self.assertEqual(actual_data.spend_step_size, 100.0)
 
@@ -1720,7 +1721,7 @@ class OptimizerAlgorithmTest(parameterized.TestCase):
       {
           'testcase_name': 'fixed_budget_scenario_with_new_budget',
           'scenario': optimizer.FixedBudgetScenario(total_budget=6000),
-          'expected_optimal_spend': np.array([1200, 1200, 1300, 1200, 1100]),
+          'expected_optimal_spend': np.array([1000, 1100, 1200, 1400, 1300]),
       },
       {
           'testcase_name': 'flexible_budget_target_roi_scenario',
@@ -1760,13 +1761,299 @@ class OptimizerAlgorithmTest(parameterized.TestCase):
     )
     spend_constraint_lower = np.array([0.5, 0.4, 0.3, 0.2, 0.1])
     spend_constraint_upper = np.array([0.5, 0.4, 0.3, 0.2, 0.1])
-    optimal_spend = self.budget_optimizer_media_and_rf.create_optimization_grid(
+    budget = (
+        scenario.total_budget
+        if isinstance(scenario, optimizer.FixedBudgetScenario)
+        else None
+    )
+    spend = self.budget_optimizer_media_and_rf.create_optimization_grid(
+        budget=budget,
+        gtol=0.01,
+    ).optimize(
+        scenario=scenario,
         spend_constraint_lower=spend_constraint_lower,
         spend_constraint_upper=spend_constraint_upper,
-        gtol=0.01,
-    ).optimize(scenario=scenario)
+    )
 
-    np.testing.assert_array_equal(optimal_spend, expected_optimal_spend)
+    np.testing.assert_array_equal(spend.optimized, expected_optimal_spend)
+
+  def test_trim_grid(self):
+    grid = optimizer.OptimizationGrid(
+        historical_spend=mock.MagicMock(),
+        use_kpi=False,
+        use_posterior=True,
+        use_optimal_frequency=False,
+        gtol=0.1,
+        round_factor=-2,
+        optimal_frequency=None,
+        selected_times=mock.MagicMock(),
+        _grid_dataset=mock.MagicMock(
+            channel=mock.MagicMock(
+                data=np.array(['ch1', 'ch2', 'ch3', 'ch4', 'ch5'])
+            ),
+            spend_grid=np.array([
+                [100, 0, 300, 0, 200],
+                [200, 100, 400, 100, 300],
+                [300, 200, 500, 200, np.nan],
+                [400, 300, np.nan, np.nan, np.nan],
+            ]),
+            incremental_outcome_grid=np.array([
+                [1.1, 0.0, 3.3, 0.0, 5.2],
+                [1.2, 2.1, 3.4, 4.1, 5.3],
+                [1.3, 2.2, 3.5, 4.2, np.nan],
+                [1.4, 2.3, np.nan, np.nan, np.nan],
+            ]),
+        ),
+    )
+    (updated_spend, updated_incremental_outcome) = grid._trim_grid(
+        spend_bound_lower=np.array([100, 100, 400, 0, 200]),
+        spend_bound_upper=np.array([400, 300, 400, 100, 300]),
+    )
+    np.testing.assert_array_equal(
+        updated_spend,
+        np.array([
+            [100, 100, 400, 0, 200],
+            [200, 200, np.nan, 100, 300],
+            [300, 300, np.nan, np.nan, np.nan],
+            [400, np.nan, np.nan, np.nan, np.nan],
+        ]),
+    )
+    np.testing.assert_array_equal(
+        updated_incremental_outcome,
+        np.array([
+            [1.1, 2.1, 3.4, 0.0, 5.2],
+            [1.2, 2.2, np.nan, 4.1, 5.3],
+            [1.3, 2.3, np.nan, np.nan, np.nan],
+            [1.4, np.nan, np.nan, np.nan, np.nan],
+        ]),
+    )
+
+  @parameterized.named_parameters(
+      {
+          'testcase_name': 'default_budget_scenario_pct_of_spend_exceeds_grid',
+          'scenario': optimizer.FixedBudgetScenario(),
+          'pct_of_spend': np.array([0.0, 0.1, 0.2, 0.3, 0.4]),
+          'spend_constraint_lower': None,
+          'spend_constraint_upper': None,
+          'error_details': (
+              'Lower bound 0 for channel ch_0 is below the mimimum spend of the'
+              ' grid 29.0.\nUpper bound 542 for channel rf_ch_0 is above the'
+              ' maximum spend of the grid 517.0.\nUpper bound 722 for channel'
+              ' rf_ch_1 is above the maximum spend of the grid 547.0.'
+          ),
+      },
+      {
+          'testcase_name': 'default_budget_scenario_constraints_exceed_grid',
+          'scenario': optimizer.FixedBudgetScenario(),
+          'pct_of_spend': None,
+          'spend_constraint_lower': np.array([1.0, 0.9, 0.8, 0.7, 0.6]),
+          'spend_constraint_upper': np.array([1.0, 0.9, 0.8, 0.7, 0.6]),
+          'error_details': (
+              'Lower bound 0 for channel ch_0 is below the mimimum spend of the'
+              ' grid 29.0.\nUpper bound 588 for channel ch_0 is above the'
+              ' maximum spend of the grid 559.0.'
+          ),
+      },
+      {
+          'testcase_name': 'fixed_budget_scenario_budget_too_large',
+          'scenario': optimizer.FixedBudgetScenario(total_budget=100_000),
+          'pct_of_spend': None,
+          'spend_constraint_lower': None,
+          'spend_constraint_upper': None,
+          'error_details': (
+              'Upper bound 27509 for channel ch_0 is above the maximum spend of'
+              ' the grid 559.0.\nUpper bound 26109 for channel ch_1 is above'
+              ' the maximum spend of the grid 530.0.\nUpper bound 23945 for'
+              ' channel ch_2 is above the maximum spend of the grid'
+              ' 486.0.\nUpper bound 25483 for channel rf_ch_0 is above the'
+              ' maximum spend of the grid 517.0.\nUpper bound 26954 for channel'
+              ' rf_ch_1 is above the maximum spend of the grid 547.0.'
+          ),
+      },
+      {
+          'testcase_name': 'fixed_budget_scenario_budget_too_small',
+          'scenario': optimizer.FixedBudgetScenario(total_budget=100),
+          'pct_of_spend': None,
+          'spend_constraint_lower': None,
+          'spend_constraint_upper': None,
+          'error_details': (
+              'Lower bound 15 for channel ch_0 is below the mimimum spend of'
+              ' the grid 29.0.\nLower bound 14 for channel ch_1 is below the'
+              ' mimimum spend of the grid 28.0.\nLower bound 13 for channel'
+              ' ch_2 is below the mimimum spend of the grid 26.0.\nLower bound'
+              ' 14 for channel rf_ch_0 is below the mimimum spend of the grid'
+              ' 27.0.\nLower bound 15 for channel rf_ch_1 is below the mimimum'
+              ' spend of the grid 29.0.'
+          ),
+      },
+      {
+          'testcase_name': (
+              'flexible_budget_target_roi_scenario_constraints_exceed_grid'
+          ),
+          'scenario': optimizer.FlexibleBudgetScenario(
+              target_metric=c.ROI, target_value=1.0
+          ),
+          'pct_of_spend': None,
+          'spend_constraint_lower': None,
+          'spend_constraint_upper': None,
+          'error_details': (
+              'Lower bound 0 for channel ch_0 is below the mimimum spend of the'
+              ' grid 29.0.\nLower bound 0 for channel ch_1 is below the mimimum'
+              ' spend of the grid 28.0.\nLower bound 0 for channel ch_2 is'
+              ' below the mimimum spend of the grid 26.0.\nLower bound 0 for'
+              ' channel rf_ch_0 is below the mimimum spend of the grid'
+              ' 27.0.\nLower bound 0 for channel rf_ch_1 is below the mimimum'
+              ' spend of the grid 29.0.\nUpper bound 588 for channel ch_0 is'
+              ' above the maximum spend of the grid 559.0.\nUpper bound 558 for'
+              ' channel ch_1 is above the maximum spend of the grid'
+              ' 530.0.\nUpper bound 512 for channel ch_2 is above the maximum'
+              ' spend of the grid 486.0.\nUpper bound 544 for channel rf_ch_0'
+              ' is above the maximum spend of the grid 517.0.\nUpper bound 576'
+              ' for channel rf_ch_1 is above the maximum spend of the grid'
+              ' 547.0.'
+          ),
+      },
+      {
+          'testcase_name': (
+              'flexible_budget_target_mroi_scenario_constraints_exceed_grid'
+          ),
+          'scenario': optimizer.FlexibleBudgetScenario(
+              target_metric=c.MROI, target_value=1.0
+          ),
+          'pct_of_spend': None,
+          'spend_constraint_lower': None,
+          'spend_constraint_upper': None,
+          'error_details': (
+              'Lower bound 0 for channel ch_0 is below the mimimum spend of the'
+              ' grid 29.0.\nLower bound 0 for channel ch_1 is below the mimimum'
+              ' spend of the grid 28.0.\nLower bound 0 for channel ch_2 is'
+              ' below the mimimum spend of the grid 26.0.\nLower bound 0 for'
+              ' channel rf_ch_0 is below the mimimum spend of the grid'
+              ' 27.0.\nLower bound 0 for channel rf_ch_1 is below the mimimum'
+              ' spend of the grid 29.0.\nUpper bound 588 for channel ch_0 is'
+              ' above the maximum spend of the grid 559.0.\nUpper bound 558 for'
+              ' channel ch_1 is above the maximum spend of the grid'
+              ' 530.0.\nUpper bound 512 for channel ch_2 is above the maximum'
+              ' spend of the grid 486.0.\nUpper bound 544 for channel rf_ch_0'
+              ' is above the maximum spend of the grid 517.0.\nUpper bound 576'
+              ' for channel rf_ch_1 is above the maximum spend of the grid'
+              ' 547.0.'
+          ),
+      },
+  )
+  def test_optimize_grid_exceeds_spend_constraint(
+      self,
+      scenario,
+      pct_of_spend,
+      spend_constraint_lower,
+      spend_constraint_upper,
+      error_details,
+  ):
+    grid = self.budget_optimizer_media_and_rf.create_optimization_grid(
+        spend_constraint_lower=0.9, spend_constraint_upper=0.9
+    )
+    error_message = (
+        'Spend allocation is not within the grid coverage:\n' + error_details
+    )
+    with self.assertRaisesWithLiteralMatch(
+        ValueError,
+        error_message,
+    ):
+      grid.optimize(
+          scenario=scenario,
+          pct_of_spend=pct_of_spend,
+          spend_constraint_lower=spend_constraint_lower,
+          spend_constraint_upper=spend_constraint_upper,
+      )
+
+  def test_optimize_grid_accuracy_raises_warning(self):
+    grid = self.budget_optimizer_media_and_rf.create_optimization_grid(
+        budget=1_000_000
+    )
+    with self.assertWarnsRegex(
+        UserWarning,
+        'Optimization accuracy may suffer owing to budget level differences.'
+        ' Consider creating a new grid with smaller `gtol` if you intend to'
+        " shrink budgets significantly. It's only a problem when you use a"
+        ' smaller budget, for which the intended step size is meant to be'
+        ' smaller for one or more channels.',
+    ):
+      grid.optimize(
+          scenario=optimizer.FixedBudgetScenario(total_budget=12),
+      )
+
+  @parameterized.named_parameters(
+      {
+          'testcase_name': 'default_fixed_budget_scenario',
+          'scenario': optimizer.FixedBudgetScenario(),
+          'fixed_budget': True,
+          'budget': None,
+          'target_roi': None,
+          'target_mroi': None,
+      },
+      {
+          'testcase_name': 'fixed_budget_scenario_with_new_budget',
+          'scenario': optimizer.FixedBudgetScenario(total_budget=6000),
+          'fixed_budget': True,
+          'budget': 6000,
+          'target_roi': None,
+          'target_mroi': None,
+      },
+      {
+          'testcase_name': 'flexible_budget_target_roi_scenario',
+          'scenario': optimizer.FlexibleBudgetScenario(
+              target_metric=c.ROI, target_value=1.0
+          ),
+          'fixed_budget': False,
+          'budget': None,
+          'target_roi': 1.0,
+          'target_mroi': None,
+      },
+      {
+          'testcase_name': 'flexible_budget_target_mroi_scenario',
+          'scenario': optimizer.FlexibleBudgetScenario(
+              target_metric=c.MROI, target_value=1.0
+          ),
+          'fixed_budget': False,
+          'budget': None,
+          'target_roi': None,
+          'target_mroi': 1.0,
+      },
+  )
+  def test_optimize_grid_matches_optimize(
+      self,
+      scenario,
+      fixed_budget,
+      budget,
+      target_roi,
+      target_mroi,
+  ):
+    spend_constraint_lower = np.array([0.5, 0.4, 0.3, 0.2, 0.1])
+    spend_constraint_upper = np.array([0.5, 0.4, 0.3, 0.2, 0.1])
+    grid = self.budget_optimizer_media_and_rf.create_optimization_grid(
+        budget=budget,
+    )
+    spend = grid.optimize(
+        scenario=scenario,
+        spend_constraint_lower=spend_constraint_lower,
+        spend_constraint_upper=spend_constraint_upper,
+    )
+    optimization_results = self.budget_optimizer_media_and_rf.optimize(
+        fixed_budget=fixed_budget,
+        budget=budget,
+        target_roi=target_roi,
+        target_mroi=target_mroi,
+        spend_constraint_lower=spend_constraint_lower,
+        spend_constraint_upper=spend_constraint_upper,
+    )
+
+    np.testing.assert_array_equal(
+        spend.optimized, optimization_results.optimized_data.spend
+    )
+    rounded_spend = np.round(spend.non_optimized, grid.round_factor).astype(int)
+    np.testing.assert_array_equal(
+        rounded_spend, optimization_results.nonoptimized_data.spend
+    )
 
   def test_optimization_grid_nans_match(self):
     self.enter_context(
@@ -1784,47 +2071,16 @@ class OptimizerAlgorithmTest(parameterized.TestCase):
     )
 
   def test_grid_search_with_target_roi_correct(self):
-    spend_grid = np.array([
-        [0, 0, 0, 0, 65900000, 40300000],
-        [100000, 100000, 100000, 100000, 66000000, 40400000],
-        [200000, 200000, 200000, 200000, 66100000, 40500000],
-        [300000, 300000, 300000, 300000, 66200000, 40600000],
-        [400000, 400000, 400000, 400000, 66300000, 40700000],
-        [500000, 500000, 500000, 500000, 66400000, 40800000],
-    ])
-    incremental_outcome_grid = np.array([
-        [0, 0, 0, 0, 31020000, 16160000],
-        [220000, 360000, 180000, 390000, 31070000, 16200000],
-        [430000, 710000, 350000, 780000, 31120000, 16240000],
-        [650000, 1060000, 530000, 1160000, 31170000, 16280000],
-        [860000, 1400000, 700000, 1520000, 31220000, 16320000],
-        [1060000, 1730000, 870000, 1880000, 31270000, 16360000],
-    ])
-
-    optimization_grid = optimizer.OptimizationGrid(
-        _grid_dataset=mock.MagicMock(
-            spend_grid=spend_grid,
-            incremental_outcome_grid=incremental_outcome_grid,
-        ),
-        historical_spend=mock.MagicMock(),
-        use_kpi=False,
-        use_posterior=True,
-        use_optimal_frequency=False,
-        round_factor=-2,
-        optimal_frequency=None,
-        selected_times=mock.MagicMock(),
+    spend = (
+        self.budget_optimizer_media_and_rf.create_optimization_grid().optimize(
+            scenario=optimizer.FlexibleBudgetScenario(
+                target_metric=c.ROI, target_value=1.0
+            )
+        )
     )
-    optimal_spend = optimization_grid.optimize(
-        scenario=optimizer.FlexibleBudgetScenario(
-            target_metric=c.ROI, target_value=1.0
-        ),
-    )
+    expected_optimal_spend = np.array([588, 558, 512, 544, 576])
 
-    expected_optimal_spend = np.array(
-        [500000, 500000, 500000, 500000, 66400000, 40300000]
-    )
-
-    np.testing.assert_array_equal(optimal_spend, expected_optimal_spend)
+    np.testing.assert_array_equal(spend.optimized, expected_optimal_spend)
 
   @mock.patch.object(analyzer.Analyzer, 'incremental_outcome', autospec=True)
   def test_optimizer_budget_with_specified_budget(
@@ -2020,6 +2276,7 @@ class OptimizerPlotsTest(absltest.TestCase):
         use_kpi=False,
         use_posterior=True,
         use_optimal_frequency=False,
+        gtol=0.1,
         round_factor=1,
         optimal_frequency=None,
         selected_times=self.meridian.expand_selected_time_dims(),
@@ -2600,6 +2857,7 @@ class OptimizerOutputTest(parameterized.TestCase):
         use_kpi=False,
         use_posterior=True,
         use_optimal_frequency=False,
+        gtol=0.1,
         round_factor=1,
         optimal_frequency=None,
         selected_times=mock.MagicMock(),